Move times related functions to minimalin.c
